refactor: Extract DefaultsTab as separate component
