+++ .trellis/p/P-implement-agent-management/e/E-ui-components/E-ui-components.md
@@ -36,7 +36,8 @@ affectedFiles:
     names with truncated descriptions
   apps/desktop/src/components/settings/agents/index.ts: Added export for
     RoleSelect component; Added ModelSelect and PersonalitySelect to barrel
-    exports
+    exports; Added DefaultsTab to barrel exports to maintain consistent import
+    patterns across the codebase.
   apps/desktop/src/components/settings/agents/__tests__/RoleSelect.test.tsx:
     Created comprehensive unit tests covering all states, functionality,
     accessibility, edge cases, and component behavior with 100% test coverage
@@ -88,11 +89,22 @@ affectedFiles:
     Removed Templates tab, mockTemplates data, TemplatesTab component,
     template-related imports, modal state properties, and openTemplateModal
     handler. Updated tab configuration to only include Library and Defaults
-    tabs. Updated JSDoc documentation to reflect two-tab navigation.
+    tabs. Updated JSDoc documentation to reflect two-tab navigation.; Removed
+    inline DefaultsTab component definition (lines 354-642). Added import for
+    DefaultsTab from './DefaultsTab'. Cleaned up unused imports that were only
+    needed by DefaultsTab (removed AgentDefaults type, useMemo, useDebounce,
+    getSliderDescription, createSliderKeyHandler, Label, Slider, Tooltip
+    components). Preserved all other functionality and imports needed by
+    remaining components.
   apps/desktop/src/components/settings/agents/__tests__/AgentsSection.test.tsx:
     Updated unit tests to reflect the new two-tab structure by removing all
     template-related test assertions and expectations. Updated test descriptions
     and assertions to match the Library and Defaults only structure.
+  apps/desktop/src/components/settings/agents/DefaultsTab.tsx:
+    Created new DefaultsTab component file with extracted code from
+    AgentsSection.tsx. Includes all configuration controls, preview panel,
+    accessibility features, and JSDoc documentation. Component maintains exact
+    same functionality as original inline version.
 log: []
 schema: v1.0
 childrenIds:
