+++ .trellis/p/P-implement-agent-management/e/E-ui-components/f/F-section-components-update/t/closed/T-extract-defaultstab-as.md
@@ -1,13 +1,33 @@
 ---
 id: T-extract-defaultstab-as
 title: Extract DefaultsTab as separate component file
-status: open
+status: done
 priority: medium
 parent: F-section-components-update
 prerequisites:
   - T-remove-templates-tab-and
-affectedFiles: {}
-log: []
+affectedFiles:
+  apps/desktop/src/components/settings/agents/DefaultsTab.tsx:
+    Created new DefaultsTab component file with extracted code from
+    AgentsSection.tsx. Includes all configuration controls, preview panel,
+    accessibility features, and JSDoc documentation. Component maintains exact
+    same functionality as original inline version.
+  apps/desktop/src/components/settings/agents/AgentsSection.tsx:
+    Removed inline DefaultsTab component definition (lines 354-642). Added
+    import for DefaultsTab from './DefaultsTab'. Cleaned up unused imports that
+    were only needed by DefaultsTab (removed AgentDefaults type, useMemo,
+    useDebounce, getSliderDescription, createSliderKeyHandler, Label, Slider,
+    Tooltip components). Preserved all other functionality and imports needed by
+    remaining components.
+  apps/desktop/src/components/settings/agents/index.ts: Added DefaultsTab to
+    barrel exports to maintain consistent import patterns across the codebase.
+log:
+  - Successfully extracted DefaultsTab component from AgentsSection.tsx into its
+    own separate file for better code organization and maintainability. All
+    functionality preserved including temperature slider (0-2), max tokens input
+    (1-4000), top P slider (0-1), reset confirmation dialog, real-time preview
+    panel, and full accessibility features. Component follows established file
+    organization patterns and maintains consistent API with parent component.
 schema: v1.0
 childrenIds: []
 created: 2025-08-19T19:26:06.035Z
