+++ apps/desktop/src/components/settings/agents/index.ts
@@ -1,5 +1,6 @@
 export { AgentCard } from "./AgentCard";
 export { AgentsSection } from "./AgentsSection";
+export { DefaultsTab } from "./DefaultsTab";
 export { EmptyLibraryState } from "./EmptyLibraryState";
 export { EmptyTemplatesState } from "./EmptyTemplatesState";
 export { ModelSelect } from "./ModelSelect";
